Mips undefined shift fix (#1011)

+++ qemu/target-mips/translate.c
@@ -34,7 +34,7 @@
 //#define MIPS_DEBUG_SIGN_EXTENSIONS
 
 /* MIPS major opcodes */
-#define MASK_OP_MAJOR(op)  (op & (0x3F << 26))
+#define MASK_OP_MAJOR(op)  (op & (((uint32_t)0x3F) << 26))
 
 enum {
     /* indirect opcode tables */
